Add a class to the catalog page and style for the titles of the columns
Add a class to the page in the editor “.catalog” and base on that class
style the titles of the columns that are h2 tags. Gave them equal
height of 80px. Style the buttons at the bottom. Need to consider if I
want all the buttons to be the same style. If so need to change the
main button css