Logo changes & Styles for the Envira Gallery
Erase the “Add support for custom header” Since the logo will be in the
image folder of the theme because we will use a SVG logo.
Erase the “Add support for custom background” because we are not going
to use that function in the theme.
Change the logo to a SVG version.
Add some styles to the Envira Gallery used in the page of Dance Crew.